backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lnys-backend"
version = "4.0.0"
description = "柠优生活大数据智能决策平台 · 后端 API"
requires-python = ">=3.10"

dependencies = [
    # Web框架
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.2.0",
    # 数据库
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "asyncmy>=0.2.9",
    "alembic>=1.13.0",
    "redis[asyncio]>=5.0.0",
    # 生产服务器
    "gunicorn>=22.0.0",
    # 限流
    "slowapi>=0.1.9",
    # ML推理
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.3.0",
    "torch>=2.2.0",
    # NLP
    "transformers>=4.38.0",
    "sentence-transformers>=3.0.0",
    "jieba>=0.42.1",
    "gensim>=4.3.0",
    # 向量数据库
    "chromadb>=0.5.0",
    # 时序
    "statsmodels>=0.14.0",
    "pmdarima>=2.0.4",
    "prophet>=1.1.5",
    # 客户分析
    "lifetimes>=0.11.3",
    "shap>=0.44.0",
    # 关联分析
    "mlxtend>=0.23.0",
    # MySQL sha256 认证
    "cryptography>=42.0.0",
    # Agent 编排 + LangGraph (v4.0)
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.2.0",
    # 可观测性 (v4.0)
    "langsmith>=0.1.0",
    # 工具
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0",  # passlib 1.7.4 不兼容 bcrypt 5.x, 和云端 4.0.1 对齐
    "sse-starlette>=1.8.0",
    # Copilot integrations
    "lark-oapi>=1.3.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
all = ["lnys-backend[dev]"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
